Update patch set 1
Patch Set 1:

Fan, I disagree. The model on the back of the device is "One A0001" - I think it is more important that we match the model name listed on the device itself. Since it includes the model number that is listed on the box, this really should be okay for the customs situation. This also matches the original request that Dave sent me - I'm assuming this came from OnePlus: "OnePlus needs a new build to install at the factory that has a model number of "One A0001" rather than the "One" that we have today."

Patch-set: 1
Reviewer: Brint E. Kriebel <116@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0