Update patch set 1
Patch Set 1:

I just want to update on my previous comment, apparently it's actually really easy to solve. 

What happened to me was changing it in the build.prop didn't change it in the Settings->Mobile Networks->Preferred network type setting. It was showing something along the lines of "Invalid Network 20 selected. Ignored" (not the exact words). I simply set it back to LTE (recommended) and it solved the problem.

I guess I assumed the situation was worse because of my prior experiences with the subject, glad it was easy to fix though!

Patch-set: 1