Update patch set 1
Patch Set 1: Code-Review+1

Ethan, the device repos are the wrong place to cater for legal issues though. If I'm not entirely mistaken, enabling support for call recording in a particular device only announces that this particular device supports call recording, but _doesn't_ override the various MCC blocks we have in the Dialer repo.

Patch-set: 1