Update patch set 8
Patch Set 8:

If you can get me a solid repro case, we can probably fix it in the sound lab. Is this with regular voice calls or on Hangouts/VOIP? There are a lot of hacks floating around that push the gains above the clipping point and also throw off the calibration :(

Also, if you do have the required tuning tools, we have the workspace file in git so you can load them right up. I can't provide the tools though, sorry.

Patch-set: 8