Update patch set 8
Patch Set 8:

np, i didnt realize what all was involved, probably wont upload another one. i dont have the tuning tool

i will try to find a consistently reproducable case though.  right now the best i have is that if you set the phone in the cup holder of my car, its horrible.  jf sounds perfect in same case.

i have tried to reproduce it by setting bacon inside a cup indoors, but it was not as bad.  the closest i have come to reproducing it indoors is laying it on a table, near a wall, with speakers/main mic facing the wall.  but thats also not nearly as bad as it is in the car cup holder.

Patch-set: 8