bacon: camera: fix memory leaks
get_parameters returns a pointer allocated from camera HAL memory with
the set of parameters. Later user space calls put_parameters to free
this pointer.

Instead, the wrapper is using strdup to return a copied pointer to user
space, so the camera pointer is never freed. Worse, it is passing this
copy to the camera HAL to free, even though the HAL never allocated it.

Instead, the wrapper should be responsible for passing the correct
pointer back to the HAL with put_parameters, and the call from user
space should be handled internally.

Change-Id: Ifdf2c61b04fbdf4700094574eb092500922f3ba8