bacon: disable 768x432 preview size
It was noticed that when showing 768x432 preview and recording a
video of larger resolution with a different aspect ratio, the preview
is distorted.  Applications should grab the largest resolution smaller than
the preferred preview size for video that matches the video resolution.  However,
bacon doesn't provide a matching preview resolution for FWVGA/WVGA
resolutions.  A typical fallback will be to the next smaller resolution than
the video.  In this case, 720x480 will work correctly.

Change-Id: I835c5f1bb141140bf83aa7ed21100c8108b66053