Update patch set 1
Patch Set 1:

That denial comes right before a kernel null pointer error and reboot.  Marc advised something like this would be better but his is for connectto and not write.

This isn't a telltale solution but it's important to note that those still having reboots on bacon are not having it with selinux set to permissive.

Patch-set: 1
Reviewer: Andrew Bartholomew <4820@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0