Update patch set 1
Patch Set 1:

Actually, the fstab generator is for generating a "normal" fstab for things that don't use fs_mgr. It won't help :(

I still think this needs to be done in a generic way, otherwise we need to deploy recovery.fstabs for every device if we want this to work.

Patch-set: 1