Update patch set 1
Patch Set 1:

I'm not sure you've tried Camera2 lately on bacon but if you take a picture with flash on Camera2 it's very easy to reproduce an issue where there's a defined line in the middle where it looks like two images were stitched together.  This only happens with flash and not in all lightning conditions.

It's not the features of Snap I'm after but basic functionality.  I like the extra features of snap, but I can work with a proper stock Camera2 as well.

Patch-set: 1