Update patch set 1
Patch Set 1:

since we speak about the reboots here, i'll share also here what i've found. When i'm using  su from CM, i get reboots randomly. In case i use SuperSU everything goes stable for hours. The testing period was for 3 days on each. while on cm su, for 3 days got around 8 reboots. When on SuperSU for 3 days, got 0 reboots. Or is something related to the su which is not beeing compiled correctly, or a permitions bug. I don't have knowledge to say what is, and am not able to catch a log... yet! Well maybe this helps out on taking conclusions as you guys are far more experienced on that than i am.

Patch-set: 1