Update patch set 1
Patch Set 1: Code-Review+1

Only +1 this because it helps improve the gps a bit, provided you flash the old libs. Old libs alone don't fix the problem 100%. However, the old libs, with this, improves the situation and increases accuracy. So again, +1 because it helps improve the situation, but this is NOT a fix.

Patch-set: 1
Label: Code-Review=+1