Update patch set 1
Patch Set 1:

Where are the logs for the reboots caused by this?  I've been running this for 5? days or so, and 0 random reboots on my bacon.  We were waiting to hear back if we would be receiving a updated modem that resolved the un-shippable blockers our testing discovered in the previous revision.

Patch-set: 1