Update patch set 2
Patch Set 2:

On my phone the compass just don't work. I've been having this problem since this version (241) was released, but never figured the modem was the problem until the previous one was introduced and the problem was gone. Now that this was brought back, the problem is back too. So I think I found the cause. :) How can I send you more information/screenshots?

Patch-set: 2