bacon: enable aec and ns
It needs to tell Android framework
this device supports aec and ns, otherwise those
effects will not be set in hal/audio_hw.c,
causing AEC and NS not enabled in
hal/msm8974/platform.c.

The libqcomvoiceprocessing.so is only telling the
framework it supports aec and ns. The real
processing is in hal/msm8974/platform.c which
set a different snd_device (endfire).

OPO-485

Change-Id: I43747ed673430ea1bbcefc4c62df08e9298f39c8