bacon: acdb: add DevicePair for handset mic + stereo headphones
If a TRS headset is plugged in, voice calls will use the handset mic
and headphone speakers. Currently, there is no DevicePair that
represents the combination of:
 TX:HANDSET_MIC_ENDFIRE_FLUENCEV5 (41) + RX:HEADSET_SPEAKER_STEREO (10)

Without this combination present audio is routed correctly, but
ACDB-LOADER would fail to send some calibration data due to the
"invalid device pair". The audio quality would therefore be
affected by whatever previous calibration data was sent, and be
noticeably different depending on the last-used audio usecase.

Correct this by adding the DevicePair mentioned above.

Issue-Id: OPO-503
Change-Id: I123d4169bed4c0acc3135a7d347aa923985a170d