Update patch set 1
Patch Set 1:

Right, because the current app is doing dumb things. It will work in the upcoming release (and won't get wedged on). C-apps have not been released on 13 yet, and if you are using a modded version of our proprietary apps then you are really on your own :)

Patch-set: 1