Update patch set 3
Patch Set 3:

The commit message is correct

The voice usecase of ACDB id 41/42 is left unmodified.  The general recording usecase of ACDB id 41/42 have been modified to remove fluence.

The general recording (VOIP) usecase of ACDB id 110/113 is identical to the voice usecase of ACDB id 41/42.

The general recording usecase of ACDB id 40 has been modified to remove the use of fluence.

The general recording usecase of ACDB id 119 has been modified to remove the use of fluence.

Patch-set: 3