Update patch set 1
Patch Set 1:

Nothing will be merged untill everything is working and ready to ship. 
(Broken camera atm as you should know since you're here yelling shipit while things are not ready).
So it's useless to make an account on gerrit and spam commits you want to see merged with "+1 Hey it does compile, no builds errors! Finally nougat for my phone! Shipit now!" and so on. This is not how things work

Patch-set: 1