Update patch set 3
Patch Set 3:

@Joe,

I'm testing this as well.

It doesn't decrypt on the first reboot for me either on a clean install of 14.1 but then I reboot and it works fine.  Essentially I'm having the same issue on a clean encryption with 14.1 as you are with an upgrade.

Please note I was using a user defined pin and it worked fine.

Patch-set: 3