Update patch set 2
Patch Set 2:

> Ethan I wouldn't be so quick to judge there - I disabled it on my
 > personal Bacon which isn't 'misconfigured' :-).
 > 
 > This commit can merge to revert the device back to status quo, but
 > there is an underlying audio issue for some % of users

Well, this property controls the selection of some of the different input/output devices. If flipping this switch is solving your issue, the underlying problem is that the devices, or their ACDB IDs is incorrect. The main reason I come out so strongly against modifying this property is because this changes about six things at once, when your issue likely lies with only one of them (DMIC Endfire/ACDB ID config).

Patch-set: 2