Update patch set 1
Patch Set 1:

> This is useless. HwUi stuffs has been removed in O.
 > 
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218733/1
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218791/1

HWUI stuff has been removed, but 

 > This is useless. HwUi stuffs has been removed in O.
 > 
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218733/1
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218791/1

Yes, that's has been removed but it is also related to dalvik stuff. Therefore 

 > This is useless. HwUi stuffs has been removed in O.
 > 
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218733/1
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218791/1

Yes, that's been removed, but it is also related to the dalvik heap which is not

 > This is useless. HwUi stuffs has been removed in O.
 > 
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218733/1
 > * https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_bacon/+/218791/1

Yes, that's been removed but it is also related to the dalvik stuff which is still in there (at least from what I know) or are the dalvik heap related values going to be removed too? In case this dalvik stuff still will be there, I'll revert it for dalvik only. But on the other hand, if it is completely useless, yes, abandon would be the option.

Patch-set: 1
Reviewer: Martin Otto <18085@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>